Renamed BGSounds to OurSounds. This class isn't confusing :D
