+++ src/main/java/com/direwolf20/buildinggadgets/common/registry/IRegistryComponent.java
@@ -1,5 +1,5 @@
 package com.direwolf20.buildinggadgets.common.registry;
 
 public interface IRegistryComponent {
-    static void setup() {}
+    void setup();
 }
